Refactor: updated imports
Fixed issue with search options

Typos fixed: Spellig mistakes
