+++ src/routes/index.ts
@@ -3,7 +3,7 @@ import addMovie from "./addMovie";
 import addtheater from "./addTheater";
 import addShow from "./addShow";
 import ticketRouter from "./tickets";
-// import {};
+import search from "./search";
 
 const router = Router();
 
@@ -11,7 +11,7 @@ router.post("/addMovie", addMovie);
 router.post("/addShow", addShow);
 router.post("/addTheater", addtheater);
 
-router.get("/search");
+router.get("/search", search);
 
 router.use("/ticket", ticketRouter);
