+++ src/routes/search/findMovies.ts
@@ -2,7 +2,7 @@ const movie = require("../../database/schemas/movie");
 
 const findMovies = async (searchString: any) => {
   try {
-    return await movie.find({ movieName: searchString });
+    return await movie.find({ movieName: { $regex: searchString } });
   } catch (error) {
     console.log(error);
   }
