+++ src/routes/tickets/bookTicket.ts
@@ -1,6 +1,6 @@
 import { RequestHandler, Request, Response } from "express";
 
-const ticket = require("../../database/schemas");
+const ticket = require("../../database/schemas/ticket");
 
 const bookTicket: RequestHandler = async (req: Request, res: Response) => {
   const ticketDetails = req.body?.ticketDetails;
