--- test/Postman Requests Collection.postman_collection.json
@@ -1,226 +1,0 @@
-{
-	"info": {
-		"_postman_id": "80871325-f784-49ed-b7c5-3efcc91bd925",
-		"name": "Postman Requests Collection",
-		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json"
-	},
-	"item": [
-		{
-			"name": "Add movie",
-			"request": {
-				"method": "POST",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "{\n    \"movieDetails\" :{\n        \"movieName\" : \"Everthing Everywhere Anytime\"\n    }\n}",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "http://localhost:8080/addMovie",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "8080",
-					"path": [
-						"addMovie"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Add show",
-			"request": {
-				"method": "POST",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "{\n    \"showDetails\" :{\n        \"showTimings\" : \"9:30 - 12:30\",\n        \"date\": \"2022-10-10\",\n        \"movies\": [\"632eb903c2e2109389a69b19\", \"632eb95ae6ef1df5afb8d649\"]\n    }\n}",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "http://localhost:8080/addShow",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "8080",
-					"path": [
-						"addShow"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Add theater",
-			"request": {
-				"method": "POST",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "{\n    \"theaterDetails\" :{\n        \"theaterName\" : \"Phoniex\",\n        \"location\": \"Pune\",\n        \"shows\": [\"632ec54d5de951cbbad4fb67\"]\n    }\n}",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "http://localhost:8080/addTheater",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "8080",
-					"path": [
-						"addTheater"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Book Ticket",
-			"request": {
-				"method": "POST",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "{\n    \"ticketDetails\" :{\n        \"theater\" : \"632ec5c05de951cbbad4fb69\",\n        \"show\": \"632ec4665a7cfc5aa6fdd9d3\"\n    }\n}",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "http://localhost:8080/bookTicket",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "8080",
-					"path": [
-						"bookTicket"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Reschedule ticket",
-			"request": {
-				"method": "POST",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "{\n    \"ticketDetails\" :{\n        \"theater\" : \"632ec5c05de951cbbad4fb69\",\n        \"show\": \"632ec4665a7cfc5aa6fdd9d3\"\n    },\n    \"id\": \"632ecb9bb67a5481b6fe18e3\"\n}",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "http://localhost:8080/rescheduleTicket",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "8080",
-					"path": [
-						"rescheduleTicket"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Search by movie name",
-			"protocolProfileBehavior": {
-				"disableBodyPruning": true
-			},
-			"request": {
-				"method": "GET",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "http://localhost:8080/search?searchString=Test&searchIn=Movie",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "8080",
-					"path": [
-						"search"
-					],
-					"query": [
-						{
-							"key": "searchString",
-							"value": "Test"
-						},
-						{
-							"key": "searchIn",
-							"value": "Movie"
-						}
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Search by theater location",
-			"protocolProfileBehavior": {
-				"disableBodyPruning": true
-			},
-			"request": {
-				"method": "GET",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": ""
-				},
-				"url": {
-					"raw": "http://localhost:8080/search?searchString=Pune&searchIn=Theater",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "8080",
-					"path": [
-						"search"
-					],
-					"query": [
-						{
-							"key": "searchString",
-							"value": "Pune"
-						},
-						{
-							"key": "searchIn",
-							"value": "Theater"
-						}
-					]
-				}
-			},
-			"response": []
-		}
-	]
-}
